tests/wpt/web-platform-tests/navigation-api/navigation-methods/traverseTo-multiple-steps.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));
  assert_equals(navigation.entries().length, 1);
  let key0 = navigation.currentEntry.key;
  await navigation.navigate("#1").committed;
  await navigation.navigate("#2").committed;
  let key2 = navigation.currentEntry.key;
  assert_equals(navigation.entries().length, 3);

  await navigation.traverseTo(key0).committed;
  assert_equals(navigation.entries().length, 3);
  assert_equals(navigation.currentEntry, navigation.entries()[0]);
  assert_equals(key0, navigation.currentEntry.key);
  await navigation.traverseTo(key2).committed;
  assert_equals(navigation.entries().length, 3);
  assert_equals(navigation.currentEntry, navigation.entries()[2]);
  assert_equals(key2, navigation.currentEntry.key);
}, "goto() can precisely traverse multiple steps in the joint session history");
</script>
