LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/traverseTo-multiple-steps-expected.txt


FAIL goto() can precisely traverse multiple steps in the joint session history assert_equals: expected 2 but got 0
